Fix Covers block
- Move attributes to a file and imported to the top of the file
- and avoid to use undeclared variables within deprecated versions
