+++ assets/src/blocks/Covers/CoversConstants.js
@@ -1,3 +1,5 @@
+const {__} = wp.i18n;
+
 export const COVERS_TYPES = {
   takeAction: 'take-action',
   campaign: 'campaign',
@@ -12,3 +14,53 @@ export const COVERS_LAYOUTS = {
 export const CAROUSEL_LAYOUT_COVERS_LIMIT = 12;
 
 export const BLOCK_NAME = 'planet4-blocks/covers';
+
+export const attributes = {
+  cover_type: {
+    type: 'string',
+    default: 'content',
+  },
+  initialRowsLimit: {
+    type: 'integer',
+    default: 1,
+  },
+  title: {
+    type: 'string',
+    default: '',
+  },
+  description: {
+    type: 'string',
+    default: '',
+  },
+  tags: {
+    type: 'array',
+    default: [],
+  },
+  post_types: {
+    type: 'array',
+    default: [],
+  },
+  posts: {
+    type: 'array',
+    default: [],
+  },
+  version: {
+    type: 'integer',
+    default: 2,
+  },
+  layout: {
+    type: 'string',
+    default: COVERS_LAYOUTS.grid,
+  },
+  isExample: {
+    type: 'boolean',
+    default: false,
+  },
+  exampleCovers: { // Used for the block's preview, which can't extract items from anything.
+    type: 'object',
+  },
+  readMoreText: {
+    type: 'string',
+    default: __('Load more', 'planet4-blocks'),
+  },
+};
